api guideline: add guideline info
