+++ app/assets/stylesheets/client_apps.scss
@@ -13,3 +13,14 @@
     color: green;
   }
 }
+
+
+#client_apps-info {
+  .image-wrapper {
+    width: 70%;
+
+    img {
+      width: 100%;
+    }
+  }
+}
