+++ app/assets/stylesheets/shared/common.scss
@@ -50,3 +50,15 @@
 .pointer {
   cursor: pointer;
 }
+
+pre {
+  word-wrap: break-word;
+  word-break: break-word;
+  padding: 10px;
+  margin: 1em 0;
+  line-height: 1.65;
+  border: 1px solid #EAEAEA;
+  color: #333;
+  background: #f8f8f8;
+  overflow: auto;
+}
