+++ app/models/event.rb
@@ -17,6 +17,7 @@
 class Event < ApplicationRecord
   include Events::Searchable
   include Events::Callback
+  include Events::TemplateField
 
   # Association
   belongs_to :event_type
@@ -56,27 +57,6 @@ class Event < ApplicationRecord
   accepts_nested_attributes_for :event_milestones, allow_destroy: true
 
   # Class Methods
-  def self.template_fields
-    fields = [
-      { code: 'event_type_name', label: 'Event Type' },
-      { code: 'creator_email', label: 'Creator' },
-      { code: 'program_name', label: 'Program' },
-      { code: 'location_name', label: 'Location Name' }
-
-    ]
-    fields.each { |field| field[:code] = "#{default_template_code}#{field[:code]}" }
-    fields += Milestone.root.fields.map { |field| { code: "#{dynamic_template_code}#{field.id}_#{field.name.downcase.split(' ').join('_')}", label: field.name } }
-    fields
-  end
-
-  def self.default_template_code
-    'de_'
-  end
-
-  def self.dynamic_template_code
-    'dy_'
-  end
-
   def self.filter(params)
     arr = keywords(params[:keyword])
     scope = all
