+++ config/locales/client_app/en.yml
@@ -16,3 +16,21 @@ en:
     edit_api_key: 'Edit API key'
     name: 'Name'
     regenerate_token: 'Regenerate token'
+    learn_more: 'Learn more'
+    need_help: 'Need help?'
+    you_can_link_to_ebs: 'You can link your data collected with EBS to other servers or services you might own. For example, enter a server URL where a json version of your data should be POSTed.'
+    api: 'API'
+    ebs_has_only_one_api: 'EBS has only one primary API for sending and retrieving data.'
+    base_url_is: 'The base URL is'
+    token: 'Token'
+    see_setup_instruction: 'Please see the instructions on how to set up service as below:'
+    click_plus_button: 'Click on plus button (+)'
+    fill_in_the_information: 'Then fill in the information as:'
+    your_given_name: 'your given name'
+    your_ip_address: 'your server ip address'
+    choose_permissions: 'choose read/write or both persmissions'
+    system_build_a_token: 'Then the system will generate a token for you to put in your server'
+    url: 'Url'
+    stand_on_own_server_and_send_request: 'By standing on your server, to test it use this command'
+    this_is_your_request_and_result: 'In this example, there is a request from client server to EBS server in order to get event types data. And it turns the following result'
+    other_endpoints: 'Other Endpoints'
