+++ config/locales/shared/en.yml
@@ -6,3 +6,4 @@ en:
     delete: 'Delete'
     cancel: 'Cancel'
     save: 'Save'
+    back: 'Back'
